docs: restructure the README and fix its snippets
- Docker and the test instructions move out of "Examples", where they were
  subsections, into a "Development" block at the end
- New section "The objects behind the client" for the classes added in 2.2.0,
  which the README did not mention although the Sugar functions are deprecated
  in their favour
- The five examples that were missing from the list are added

Fixes three defects in the code samples:

- The account_objects sample called getBody() on the return of syncRequest(),
  which is a BaseResponse and has no such method, so it died with a fatal error.
  It also carried an "Account Info" heading over an AccountObjectsRequest.
- The payment sample derived the operational wallet from the standby seed.
- It sent 100 XRP, exactly what the faucet hands out, so following it with a
  faucet wallet yields tecUNFUNDED_PAYMENT because reserve and fee are left
  uncovered.

The payment sample now leads with submitAndWait($tx, autofill: true,
wallet: $wallet), which only started working with the getSignedTx() fix in
2.2.0, and keeps the sign-then-submit form as the alternative the files in
examples/ use. All samples were run against the Testnet.

+++ src/Core/RippleAddressCodec/BaseX.php
@@ -84,7 +84,7 @@ class BaseX
             $i = 0;
             for ($it1 = $size - 1; ($carry !== 0 || $i < $length) && ($it1 !== -1); $it1--, $i++) {
                 $carry += $this->unsignedRightShift((int)abs(256 * $b58[$it1]), 0);
-                $b58[$it1] = $this->unsignedRightShift((int)abs($carry % $this->base), 0);
+                $b58[$it1] = $this->unsignedRightShift(abs($carry % $this->base), 0);
                 $carry = $this->unsignedRightShift((int)abs($carry / $this->base), 0);
             }
             if ($carry !== 0) {
